feat: 걸린 시간 소수점 세자리까지 출력되도록 수정

+++ onetofifty/src/app/Utils/handleTime.tsx
@@ -4,7 +4,7 @@ const getCurrentTime = () => {
 };
 
 const getElapsedTime = (array: number[]) => {
-  const time: number = parseInt((array[1] - array[0]).toFixed(3));
+  const time: number = parseFloat((array[1] - array[0]).toFixed(3));
   return time;
 };
